+++ src/components/simple/ArrowRight.js
